순서변경 구현중

+++ src/main/java/kr/taeu/handa/domain/todoItem/dto/ModifyOrderRequest.java
@@ -14,10 +14,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ModifyOrderRequest {
 	@NotNull
-	private Integer order;
+	private Double order;
 	
 	@Builder
-	public ModifyOrderRequest(@NonNull Integer order) {
+	public ModifyOrderRequest(@NonNull Double order) {
 		this.order = order;
 	}
